fix: Move parsing to AbilityTreeDump
